Fixed a bug that could lead to missing newlines in response emails when no content was provided
git-svn-id: https://plugins.svn.wordpress.org/planso-forms/trunk@1258258 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ includes/submit_notifications.php
@@ -366,15 +366,30 @@
 	if(!isset($j->clean_attachments) || $j->clean_attachments===true || $j->clean_attachments===1){
 		if($has_attachments && count($attachments)>0){
 			foreach($attachments as $f){
-				chmod($f,0777);
-				unlink($f);
+				if(is_array($f)){
+					foreach($f as $ff){
+						if(is_file($ff)){
+							chmod($ff,0777);
+							unlink($ff);
+						}
+					}
+				} else {
+					if(is_file($f)){
+						chmod($f,0777);
+						unlink($f);
+					}
+				}
 			}
-			chmod($upload_dir,0777);
-			$files = array_diff(scandir($upload_dir), array('.','..')); 
-	    foreach($files as $file){
-	      unlink($upload_dir.'/'.$file); 
-	    } 
-			rmdir($upload_dir);
+			if(is_dir($upload_dir)){
+				chmod($upload_dir,0777);
+				$files = array_diff(scandir($upload_dir), array('.','..')); 
+		    foreach($files as $file){
+		    	if(is_file($file)){
+		      	unlink($upload_dir.'/'.$file); 
+		      }
+		    } 
+				rmdir($upload_dir);
+			}
 		}
 	}
 	if(!isset($affiliate_params)){
